+++ nbinteract/cli.py
@@ -1,16 +1,20 @@
-#!/usr/bin/env python
 '''Converts notebooks to interactive HTML pages or Gitbook pages.
 
 Usage:
-  nbinteract SPEC NOTEBOOKS ...
-  nbinteract [options] SPEC NOTEBOOKS ...
+  nbinteract init
+  nbinteract NOTEBOOKS ...
+  nbinteract [options] NOTEBOOKS ...
   nbinteract (-h | --help)
 
-Arguments:
-  SPEC       BinderHub spec for Jupyter image. Must be in the format:
-             `{username}/{repo}/{branch}`. For example:
-             'SamLau95/nbinteract-image/master'.
+`nbinteract init` initializes a GitHub project for nbinteract. It
+provides guided help to set up a requirements.txt file (if needed) and a Binder
+image for the project.
 
+`nbinteract NOTEBOOKS ...` converts notebooks into HTML pages. Note that
+running this command outside a GitHub project initialized with `nbinteract
+init` requires you to specify the --spec SPEC option.
+
+Arguments:
   NOTEBOOKS  List of notebooks or folders to convert. If folders are passed in,
              all the notebooks in each folder are converted. The resulting HTML
              files are created adjacent to their originating notebooks and will
@@ -21,6 +25,11 @@ Arguments:
 
 Options:
   -h --help                  Show this screen
+  -s SPEC --spec SPEC        BinderHub spec for Jupyter image. Must be in the
+                             format: `{username}/{repo}/{branch}`. For example:
+                             'SamLau95/nbinteract-image/master'. This flag is
+                             **required** unless a .nbinteract.json file exists
+                             in the project root with the "spec" key.
   -t TYPE --template TYPE    Specifies the type of HTML page to generate. Valid
                              types: full (standalone page), partial (embeddable
                              page), or gitbook (embeddable page for GitBook).
@@ -34,11 +43,21 @@ Options:
   -i FOLDER --images=FOLDER  Extracts images from HTML and writes into FOLDER
                              instead of encoding images in base64 in the HTML.
                              Requires -o option to be set as well.
+  -e --execute               Executes the notebook before converting to HTML,
+                             functioning like the equivalent flag for
+                             nbconvert. Configure NbiExecutePreprocessor to
+                             change conversion instead of the base
+                             ExecutePreprocessor.
 '''
 from docopt import docopt, DocoptExit
 from glob import glob
 import os
 import re
+import sys
+from textwrap import wrap
+import subprocess
+import json
+import fnmatch
 
 import nbformat
 from traitlets.config import Config
@@ -48,6 +67,8 @@ BLUE = "\033[0;34m"
 RED = "\033[91m"
 NOCOLOR = "\033[0m"
 
+CONFIG_FILE = '.nbinteract.json'
+
 VALID_TEMPLATES = set(['full', 'gitbook', 'partial', 'local'])
 
 SPEC_REGEX = re.compile('\S+/\S+/\S+')
@@ -56,6 +77,26 @@ SPEC_REGEX = re.compile('\S+/\S+/\S+')
 # parse them properly
 CLOSING_DIV_REGEX = re.compile('\s+</div>')
 
+BINDER_BASE_URL = 'https://mybinder.org/v2/gh/'
+REQUIREMENTS_DOCS = 'http://mybinder.readthedocs.io/en/latest/using.html#id8'
+DOCKER_DOCS = 'https://mybinder.readthedocs.io/en/latest/dockerfile.html'
+
+ERROR = 1
+SUCCESS = 0
+
+
+def binder_spec_from_github_url(github_url):
+    """
+    Converts GitHub origin into a Binder spec.
+
+    For example:
+    git@github.com:SamLau95/nbinteract.git -> SamLau95/nbinteract/master
+    https://github.com/Calebs97/riemann_book -> Calebs97/riemann_book/master
+    """
+    tokens = re.split(r'\W+', github_url.replace('.git', ''))
+    # The username and reponame are the last two tokens
+    return '{}/{}/master'.format(tokens[-2], tokens[-1])
+
 
 def flatmap(fn, iterable, *args, **kwargs):
     return [
@@ -63,12 +104,54 @@ def flatmap(fn, iterable, *args, **kwargs):
     ]
 
 
-def log(text):
-    print('{}[nbinteract]{} {}'.format(BLUE, NOCOLOR, text))
+def color(text, text_color):
+    return text_color + text + NOCOLOR
 
 
-def error(text):
-    print('{}[nbinteract]{} {}'.format(RED, NOCOLOR, text))
+def log(text='', line_length=80, heading='[nbinteract] ', text_color=BLUE):
+    width = line_length - len(heading)
+    for line in wrap(text, width, subsequent_indent='  ') or ['']:
+        print(color(heading, text_color) + line)
+
+
+def error(text='', line_length=80, heading='[nbinteract] '):
+    log(text, line_length, heading, text_color=RED)
+
+
+def yes_or_no(question, default="yes"):
+    """Ask a yes/no question via input() and return their answer.
+
+    "question" is a string that is presented to the user.
+    "default" is the presumed answer if the user just hits <Enter>.
+        It must be "yes" (the default), "no" or None (meaning
+        an answer is required of the user).
+
+    The "answer" return value is True for "yes" or False for "no".
+    """
+    valid = {"yes": True, "y": True, "ye": True, "no": False, "n": False}
+    if default is None:
+        prompt = " [y/n] "
+    elif default == "yes":
+        prompt = " [Y/n] "
+    elif default == "no":
+        prompt = " [y/N] "
+    else:
+        raise ValueError("invalid default answer: '%s'" % default)
+
+    while True:
+        sys.stdout.write(
+            '{}[nbinteract]{} {}{}'.format(BLUE, NOCOLOR, question, prompt)
+        )
+        choice = input().lower()
+        if default is not None and choice == '':
+            return valid[default]
+        elif choice in valid:
+            return valid[choice]
+        else:
+            sys.stdout.write(
+                "Please respond with 'yes' or 'no' "
+                "(or 'y' or 'n').\n"
+            )
 
 
 def main():
@@ -76,6 +159,23 @@ def main():
     Parses command line options and runs nbinteract.
     """
     arguments = docopt(__doc__)
+    if arguments['init']:
+        return_code = init()
+        sys.exit(return_code)
+
+    run_converter(arguments)
+
+
+def run_converter(arguments):
+    """
+    Converts notebooks to HTML files. Returns list of output file paths
+    """
+    # Get spec from config file
+    if os.path.isfile(CONFIG_FILE):
+        with open(CONFIG_FILE) as f:
+            config = json.load(f)
+            arguments['--spec'] = arguments['--spec'] or config['spec']
+
     check_arguments(arguments)
 
     notebooks = flatmap(
@@ -86,13 +186,15 @@ def main():
 
     exporter = init_exporter(
         extract_images=arguments['--images'],
-        spec=arguments['SPEC'],
+        spec=arguments['--spec'],
         template_file=arguments['--template'],
-        button_at_top=(not arguments['--no-top-button'])
+        button_at_top=(not arguments['--no-top-button']),
+        execute=arguments['--execute'],
     )
 
     log('Converting notebooks to HTML...')
 
+    output_files = []
     for notebook in notebooks:
         output_file = convert(
             notebook,
@@ -100,6 +202,7 @@ def main():
             output_folder=arguments['--output'],
             images_folder=arguments['--images']
         )
+        output_files.append(output_file)
         log('Converted {} to {}'.format(notebook, output_file))
 
     log('Done!')
@@ -107,8 +210,133 @@ def main():
     if arguments['--images']:
         log('Resulting images located in {}'.format(arguments['--images']))
 
+    return output_files
+
+
+def init():
+    '''
+    Initializes git repo for nbinteract.
+
+    1. Checks for requirements.txt or Dockerfile, offering to create a
+       requirements.txt if needed.
+    2. Sets the Binder spec using the `origin` git remote in .nbinteract.json.
+    3. Prints a Binder URL so the user can debug their image if needed.
+    '''
+    log('Initializing folder for nbinteract.')
+    log()
+
+    log('Checking to see if this folder is the root folder of a git project.')
+    if os.path.isdir('.git'):
+        log("Looks like we're in the root of a git project.")
+    else:
+        error(
+            "This folder doesn't look like the root of a git project. "
+            "Please rerun nbinteract init in the top-level folder of a "
+            "git project."
+        )
+        return ERROR
+    log()
+
+    log('Checking for requirements.txt or Dockerfile.')
+    if os.path.isfile('Dockerfile'):
+        log(
+            'Dockerfile found. Note that Binder will use the Dockerfile '
+            'instead of the requirements.txt file, so you should make sure '
+            'your Dockerfile follows the format in {docker_docs}'
+            .format(docker_docs=DOCKER_DOCS)
+        )
+    elif os.path.isfile('requirements.txt'):
+        log('requirements.txt found.')
+    else:
+        log('No requirements.txt file found.')
+        yes = yes_or_no(
+            'Would you like to create a sample requirements.txt file?'
+        )
+        if yes:
+            # TODO(sam): Don't hard-code requirements.txt
+            with open('requirements.txt', 'w') as f:
+                f.writelines(['numpy\n', 'nbinteract\n'])
+            log(
+                'Created requirements.txt. Edit this file now to include the '
+                'rest of your dependencies, then rerun nbinteract init.'
+            )
+            return SUCCESS
+        else:
+            log(
+                'Please manually create a requirements.txt file, then rerun '
+                'nbinteract init.'
+            )
+            return SUCCESS
+    log()
+
+    log('Generating .nbinteract.json file...')
+    if os.path.isfile(CONFIG_FILE):
+        log(
+            ".nbinteract.json already exists, skipping generation. If you'd "
+            "like to regenerate the file, remove .nbinteract.json and rerun "
+            "this command."
+        )
+        log()
+        log("Initialization success!")
+        return SUCCESS
+
+    try:
+        github_origin = str(
+            subprocess.check_output(
+                'git remote get-url origin',
+                stderr=subprocess.STDOUT,
+                shell=True
+            ), 'utf-8'
+        ).strip()
+    except subprocess.CalledProcessError as e:
+        error(
+            "No git remote called origin found. Please set up your project's"
+            "origin remote to point to a GitHub URL.\ngit error: {}".format(e)
+        )
+        return ERROR
+
+    if 'github' not in github_origin:
+        error(
+            "Your project's origin remote {} doesn't look like a github "
+            "URL. This may cause issues with Binder, so please double check "
+            "your .nbinteract.json file after this script finishes. "
+            "Continuing as planned..."
+        )
+
+    binder_spec = binder_spec_from_github_url(github_origin)
+    with open(CONFIG_FILE, 'w') as f:
+        json.dump({'spec': binder_spec}, f, indent=4)
+    log('Created .nbinteract.json file successfully')
+    log()
+
+    log(
+        'Initialization complete! Now, you should make a git commit with the '
+        'files created by in this process and push your commits to GitHub.'
+    )
+    log()
+    log(
+        'After you push, you should visit {} and verify that your Binder '
+        'image successfully starts.'.format(BINDER_BASE_URL + binder_spec)
+    )
+
 
 def check_arguments(arguments):
+    if not arguments['--spec']:
+        error(
+            '--spec flag not set and no .nbinteract.json file found. Rerun '
+            'this command with the --spec flag or run `nbinteract init` to '
+            'resolve this issue.'
+        )
+        raise DocoptExit()
+
+    if not SPEC_REGEX.match(arguments['--spec']):
+        error(
+            'Spec must be in the format {username}/{repo}/{branch} but got ' +
+            arguments['--spec'] + '.\n'
+            'Exiting...'
+        )
+        raise DocoptExit()
+
     if arguments['--images'] and not arguments['--output']:
         error(
             'If --images is specified, --output must also be specified. '
@@ -123,13 +351,6 @@ def check_arguments(arguments):
         )
         raise DocoptExit()
 
-    if not SPEC_REGEX.match(arguments['SPEC']):
-        error(
-            'Spec must be in the format {username}/{repo}/{branch}. '
-            'Exiting...'
-        )
-        raise DocoptExit()
-
 
 def expand_folder(notebook_or_folder, recursive=False):
     """
@@ -138,29 +359,49 @@ def expand_folder(notebook_or_folder, recursive=False):
 
     If recursive is True, recurses into subdirectories.
     """
-    if os.path.isfile(notebook_or_folder):
-        return [notebook_or_folder]
-    elif os.path.isdir(notebook_or_folder):
-        matcher = '{}/**/*.ipynb' if recursive else '{}/*.ipynb'
-        return glob(matcher.format(notebook_or_folder), recursive=recursive)
-    else:
+    is_file = os.path.isfile(notebook_or_folder)
+    is_dir = os.path.isdir(notebook_or_folder)
+    if not (is_file or is_dir):
         raise ValueError(
             '{} is neither an existing file nor a folder.'
             .format(notebook_or_folder)
         )
 
+    if is_file:
+        return [notebook_or_folder]
 
-def init_exporter(extract_images, **exporter_config):
+    # Now we know the input is a directory
+    if not recursive:
+        return glob('{}/*.ipynb'.format(notebook_or_folder))
+
+    # Recursive case
+    return [
+        os.path.join(folder, filename)
+        for folder, _, filenames in os.walk(notebook_or_folder)
+        # Skip folders that start with .
+        if not os.path.basename(folder).startswith('.')
+        for filename in fnmatch.filter(filenames, '*.ipynb')
+    ]
+
+
+def init_exporter(extract_images, execute, **exporter_config):
     """
     Returns an initialized exporter.
     """
     config = Config(InteractExporter=exporter_config)
 
+    preprocessors = []
     if extract_images:
         # Use ExtractOutputPreprocessor to extract the images to separate files
-        config.InteractExporter.preprocessors = [
-            'nbconvert.preprocessors.ExtractOutputPreprocessor',
-        ]
+        preprocessors.append(
+            'nbconvert.preprocessors.ExtractOutputPreprocessor'
+        )
+    if execute:
+        # Use the NbiExecutePreprocessor to correctly generate widget output
+        # for interact() calls.
+        preprocessors.append('nbinteract.preprocessors.NbiExecutePreprocessor')
+
+    config.InteractExporter.preprocessors = preprocessors
 
     exporter = InteractExporter(config=config)
     return exporter